requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
distro==1.8.0
frozenlist==1.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gspread==5.11.0
gspread-dataframe==3.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
netaddr==0.8.0
numpy==1.25.2
oauthlib==3.2.2
outline-vpn-api==3.0.0
pandas==2.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
yookassa==2.4.0
